Refactor: add model group #13
